% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{ClusterVariable}
\alias{ClusterVariable}
\title{ClusterVariable}
\usage{
ClusterVariable(corpus, resHCPC)
}
\arguments{
\item{corpus}{Corpus.}

\item{resHCPC}{a clustering from StartClus.}
}
\value{
a corpus.
}
\description{
This function import cluster data into your corpus.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
res <- StartCA(dtm,ncp=5)
resHCPC <- StartClus(res,clus=3)
newcorpus <- ClusterVariable(corpus, resHCPC)
}
\author{
Bouchet-Valat Milan
}
