\name{copyVarFreq}
\alias{copyVarFreq}
\title{Copy variable frequencies table to clipboard}
\description{Copy a meta-data variables frequencies table to the clipboard.}
\details{This command copies a HTML version of a previously computed meta-data
         variables frequencies table to the clipboard, to allow pasting it into e.g.
         a word processor.
}
\seealso{\code{\link{varTableDlg}}, \code{\link{varCrossTableDlg}}, \code{\link{HTML2clip}} }
