\name{copyVocabulary}
\alias{copyVocabulary}
\title{Copy vocabulary table to clipboard}
\description{Copy a vocabulary table to the clipboard.}
\details{This command copies a HTML version of a previously computed vocabulary
         table to the clipboard, to allow pasting it into e.g. a word processor.
}
\seealso{\code{\link{varVocabularyDlg}}, \code{\link{docVocabularyDlg}}, \code{\link{HTML2clip}} }
