% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAAgr.R
\name{CorpusCa}
\alias{CorpusCa}
\title{CorpusCa}
\usage{
CorpusCa(corpus, dtm = NULL, variables = NULL, sparsity = 0.9, ...)
}
\arguments{
\item{corpus}{corpus}

\item{dtm}{Document-term matrix}

\item{variables}{variables.}

\item{sparsity}{numeric argument that indicate the sparsity of your dtm.}

\item{...}{others arguments}
}
\value{
a correspondence analysis
}
\description{
This function make a correspondence analysis, you should observe it by explor.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
res <- CorpusCa(corpus,dtm)
}
\author{
Bouchet-Valat Milan
}
