% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.R
\name{CreateCloud}
\alias{CreateCloud}
\title{CreateCloud}
\usage{
CreateCloud(corpus, language, max.words = 50, ...)
}
\arguments{
\item{corpus}{corpus.}

\item{language}{numeric argument that indicate the sparsity of your dtm.}

\item{max.words}{numeric argument that indicate the maximum of words used.}

\item{...}{others arguments}
}
\value{
a wordcloud
}
\description{
Create wordcloud
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
cloud <- CreateCloud(corpus,"fr")
}
\author{
Bouchet-Valat Milan
}
