% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creadictio.R
\name{CreateDictionary}
\alias{CreateDictionary}
\title{CreateDictionary}
\usage{
CreateDictionary(corpus, dtm, language)
}
\arguments{
\item{corpus}{corpus.}

\item{dtm}{a document-term matrix.}

\item{language}{language of your corpus.}
}
\value{
a dictionary: matrix with frequencies , stems, meaningless words
}
\description{
This function create a dictionary with the words of your corpus.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
dictionary <- CreateDictionary(corpus,dtm,"fr")
}
\author{
Bouchet-Valat Milan
}
