% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creadtm.R
\name{Createdtm}
\alias{Createdtm}
\title{Createdtm}
\usage{
Createdtm(corpus, sparsity = 0.98)
}
\arguments{
\item{corpus}{corpus.}

\item{sparsity}{numeric argument that indicate the sparsity of your dtm.}
}
\value{
a matrix call dtm.
}
\description{
This function create a document-term matrix (dtm).
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
}
\author{
Bouchet-Valat Milan
}
