% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphtree.R
\name{CreateTree}
\alias{CreateTree}
\title{CreateTree}
\usage{
CreateTree(dtm, freq = 0, n = 100)
}
\arguments{
\item{dtm}{document term matrix.}

\item{freq}{term minimum frequency.}

\item{n}{nomber of term.}
}
\value{
a word graph
}
\description{
This function create a word graph
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
tree <- CreateTree(dtm,freq=3,n=100)
}
\author{
Bouchet-Valat Milan
}
