% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coocurence.R
\name{Freqterm}
\alias{Freqterm}
\title{Freqterm}
\usage{
Freqterm(dtm, variable = NA, nb = 10)
}
\arguments{
\item{dtm}{a document-term matrix.}

\item{variable}{variable.}

\item{nb}{number of term.}
}
\value{
an array of frequencies.
}
\description{
This function return the freq of each term regarding variables.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
freq <- Freqterm(dtm,nb=20)
}
\author{
Bouchet-Valat Milan
}
