% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coocurence.R
\name{LexicalSummary}
\alias{LexicalSummary}
\title{LexicalSummary}
\usage{
LexicalSummary(corpus, dtm, variable = NULL, unit = c("document", "global"))
}
\arguments{
\item{corpus}{a corpus.}

\item{dtm}{a document-term matrix.}

\item{unit}{vector.}
}
\value{
summary of each document in the corpus.
}
\description{
This function sumarise lexical information from your corpus.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
Sumlex <- LexicalSummary(corpus,dtm)
}
\author{
Bouchet-Valat Milan
}
