% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicmodel.R
\name{Plottopic}
\alias{Plottopic}
\title{PLottopic}
\usage{
Plottopic(restopic)
}
\arguments{
\item{restopic}{a topic model.}
}
\value{
a graphic.
}
\description{
This function plot data from your topic model.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
res <- Topicmodel(dtm)
Plottopic(res)
}
\author{
Bouchet-Valat Milan
}
