% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_final.r
\name{setCorpusVariables}
\alias{setCorpusVariables}
\title{setCorpusVariables}
\usage{
setCorpusVariables(corpus, dset)
}
\arguments{
\item{corpus}{your corpus.}

\item{dset}{data set.}
}
\value{
a corpus with new variables
}
\description{
Set corpus meta-data variables from the active data set.
}
\examples{

corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dset <- data.frame(x=c(1:length(corpus)))
corpus <- setCorpusVariables(corpus,dset)
}
\author{
Bouchet-Valat Milan
}
