% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creadictio.R
\name{SortDictionary}
\alias{SortDictionary}
\title{SortDictionary}
\usage{
SortDictionary(dictionary, by = "Occurrences")
}
\arguments{
\item{dictionary}{corpus}

\item{by}{condition}
}
\value{
a sorted dictionary
}
\description{
This function order your dictionary.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
dictionary <- CreateDictionary(corpus,dtm,"fr")
dictionarysort <- SortDictionary(dictionary,"alphabetic")
}
\author{
Bouchet-Valat Milan
}
