% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_final.r
\name{SplitCorpus}
\alias{SplitCorpus}
\title{SplitCorpus}
\usage{
SplitCorpus(corpus, nb)
}
\arguments{
\item{corpus}{your corpus.}

\item{nb}{number of paragraph you want.}
}
\value{
a splited corpus.
}
\description{
This function split your corpus.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
splitedcorpus <- SplitCorpus(corpus,3)
}
\author{
Bouchet-Valat Milan
}
