% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{StartCA}
\alias{StartCA}
\title{StartCA}
\usage{
StartCA(dtm, ncp = 5)
}
\arguments{
\item{dtm}{document term matrix}

\item{ncp}{nomber of axes you want to keep}
}
\value{
a correspondence analysis
}
\description{
This function make a correspondence analysis, you should observe it by explor.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
res <- StartCA(dtm,ncp=5)
}
\author{
Bouchet-Valat Milan
}
