% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA.R
\name{StartClus}
\alias{StartClus}
\title{StartClus}
\usage{
StartClus(res, clus = 0)
}
\arguments{
\item{res}{result of your correspondence analysis.}

\item{clus}{nomber of group you want to keep. Keep 0 in order to manualy cut the tree.}
}
\value{
a clustering.
}
\description{
This function make a clustering.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
res <- StartCA(dtm,ncp=5)
resHCPC <- StartClus(res,clus=3)
}
\author{
Bouchet-Valat Milan
}
