% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creadictio.R
\name{StemDtm}
\alias{StemDtm}
\title{StemDtm}
\usage{
StemDtm(dtm, dictionary)
}
\arguments{
\item{dtm}{a document-term matrix.}

\item{dictionary}{dictionary.}
}
\value{
a document-term matrix composed of stem.
}
\description{
This function change words by stems.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
dictionary <- CreateDictionary(corpus,dtm,"fr")
newdtm <- StemDtm(dtm,dictionary)
}
\author{
Bouchet-Valat Milan
}
