% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{terms_graph}
\alias{terms_graph}
\title{terms_graph}
\usage{
terms_graph(dtm, n = 100, min_occ = 0)
}
\arguments{
\item{dtm}{A \code{DocumentTermMatrix} object.}

\item{n}{The maximum number of terms to represent.}

\item{min_occ}{The minimum number of occurrences for a term to be retained.}
}
\value{
The ID of the plot returned by \code{\link{tkplot}}.
}
\description{
Plot a graph of terms.
}
\examples{

file <- system.file("texts", "reut21578-factiva.xml", package="tm.plugin.factiva")
corpus <- import_corpus(file, "factiva", language="en")
dtm <- build_dtm(corpus)
terms_graph(dtm, 100, 3)

}
