% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicmodel.R
\name{Topicmodel}
\alias{Topicmodel}
\title{Topicmodel}
\usage{
Topicmodel(dtm)
}
\arguments{
\item{dtm}{document term matrix.}
}
\description{
This function create a topic model.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
res <- Topicmodel(dtm)
}
\author{
Bouchet-Valat Milan
}
