% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri1var.R
\name{tri1var}
\alias{tri1var}
\title{Tri1var}
\usage{
tri1var(corpus, var)
}
\arguments{
\item{corpus}{a document-term matrix.}

\item{var}{variable.}
}
\value{
plot a barchart.
}
\description{
This function make an univariate analyse.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
tri1var(corpus,"reg")
}
\author{
Bouchet-Valat Milan
}
