% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri1var.R
\name{tri2var}
\alias{tri2var}
\title{Tri2var}
\usage{
tri2var(corpus, var.row, var.col)
}
\arguments{
\item{corpus}{a document-term matrix.}

\item{var.row}{variable.}

\item{var.col}{variable.}
}
\value{
plot a barchart.
}
\description{
This function make an bivariate analyse.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
tri2var(corpus,"reg","statut")
}
\author{
Bouchet-Valat Milan
}
