% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicmodel.R
\name{Viewtopic}
\alias{Viewtopic}
\title{Viewtopic}
\usage{
Viewtopic(restopic)
}
\arguments{
\item{restopic}{a topic model.}
}
\value{
an array.
}
\description{
This function create resume of your topic model.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
dtm <- Createdtm(corpus=corpus,sparsity=0.98)
res <- Topicmodel(dtm)
tab <- Viewtopic(res)
}
\author{
Bouchet-Valat Milan
}
