% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_final.r
\name{ImportCorpus}
\alias{ImportCorpus}
\title{ImportCorpus}
\usage{
ImportCorpus(file, format, language = "fr", encoding = "UTF-8")
}
\arguments{
\item{file}{file path you want to import.}

\item{format}{file format.}

\item{language}{language of your corpus.}

\item{encoding}{encoding of the file.}
}
\value{
a corpus.
}
\description{
This function import your corpus.
}
\examples{
corpus <- ImportCorpus(system.file("texts/Corpus_final.txt",package = "functiontest"),"alceste","fr")
}
\author{
Bouchet-Valat Milan
}
